Moved spotify data models to separate package
