+++ data_processing/fetch/spotify_api/data_models/spotify_album_tracks_model.py
