+++ data_processing/fetch/spotify_api/data_models/spotify_search_album_model.py
